update fedavg (use all train data)

+++ data/dataloaders/femnist.py
@@ -25,7 +25,7 @@ class FemnistDataset(Dataset[XY]):
 
     def __getitem__(self, idx: int) -> XY:
         imgs = torch.tensor(self.imgs[idx])
-        labels = torch.tensor(self.labels[idx])
+        labels = torch.tensor(self.labels[idx], dtype=torch.long)
         return imgs, labels
 
 def get_loader(path_to_pickle, batch_size=32, shuffle=True):
